+++ protected/humhub/docs/CHANGELOG.md
@@ -23,8 +23,9 @@ Important note for LDAP users: There is a new setting "ID Attribute" which shoul
 - Fix: trigger richtext `clear` when submitting comment.
 - Fix: missing return in `FileContent::beforeValidate`
 - Fix: Mentioning search with `-` not working
-- Fix #2730 Mentioning search with `-` not working
-- Fix File search with suffix not working
+- Fix #2730: Mentioning search with `-` not working
+- Fix: File search with suffix not working
+- Enh: Added SearchAttributesEvent to improve content addon indexing (comment/file)
 
 1.2.2  (August 2, 2017)
 --------------------------------